Fix deadlock on PlanExecutor.
The deadlock happened when the Main thread was blocked, synchronously
waiting  for the `RunningOperationEndedEvent` to be signaled, which
in turn is only signaled after the entire operation is complete.
This completion only happens after the file transfer task is complete,
which itself has various threads running and synchronously invoking
callbacks via the Main thread's `Dispatcher` to update the progress
on the UI/Console. This `Dispatcher` however, was currently blocked
waiting for the aforementioned event.

How do we solve this?

We make the Main thread not really block by specifying a timeout when
checking whether the event was set or not, giving us the chance to
tell the dispatcher to process more events in the meantime.