`BackupPlanFile`s & `BackupedFile`s no longer reference `BackupPlan`. Now they reference a `StorageAccount`.
The reason for this is that a `SyncOperation` should download a list
of all files that are in the remote storage and save their infos in
the local database. Then during the creation of a `RestorePlan`, the
tree view needs to show files that have already been backuped, and also
show files that have been synchronized - which are not local.
To make it easier, we create/update `BackupPlanFile`s for each sync'd
file. That would then require us to link new files to an existing
`BackupPlanFile` - which might have not been created yet (think about
backuping in Machine A then moving to Machine B to restore it & continue
backuping).
What should we do? We link the files to an `StorageAccount` instead of
to a `BackupPlan`.